+++ catalog/app/containers/Bucket/PackageEditor/FilesWorkspace.tsx
